+++ collaboredit/package.json
@@ -2,10 +2,12 @@
   "name": "collaboredit",
   "version": "0.0.1",
   "description": "A simple application that was written as a reasearch framework for CRDTs.",
+  "module": "ES2015",
   "main": "index.js",
   "scripts": {
-    "start": "webpack serve",
-    "build": "webpack --config webpack.config.js",
+   "start": "webpack serve --config webpack.config.js" ,
+    "start:server": "./node_modules/bin/ts-node ./src/rtc/index.ts",
+    "build": "webpack -config webpack.config.js && webpack --config webpack.server.config.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
@@ -21,10 +23,12 @@
   "dependencies": {
     "@babel/plugin-transform-runtime": "^7.12.1",
     "@rebass/forms": "^4.0.6",
+    "@types/express": "^4.17.9",
     "@types/jest": "^26.0.15",
     "@types/node": "^14.14.9",
     "@types/react": "^17.0.0",
     "@types/react-dom": "^17.0.0",
+    "@types/react-router-dom": "^5.1.6",
     "@types/rebass": "^4.0.7",
     "@types/ws": "^7.4.0",
     "automerge": "^0.14.1",
@@ -33,6 +37,7 @@
     "react-dom": "^17.0.1",
     "react-router-dom": "^5.2.0",
     "rebass": "^4.0.7",
+    "webpack-node-externals": "^2.5.2",
     "ws": "^7.4.0"
   },
   "devDependencies": {
@@ -40,12 +45,16 @@
     "@babel/preset-env": "^7.12.7",
     "@babel/preset-react": "^7.12.7",
     "@babel/preset-typescript": "^7.12.7",
+    "@types/jest-diff": "^24.3.0",
     "@types/rebass__forms": "^4.0.4",
     "@types/webpack": "^4.41.25",
     "@types/webpack-dev-server": "^3.11.1",
     "babel-loader": "^8.2.1",
     "html-webpack-plugin": "^4.5.0",
+    "jest-diff": "^26.6.2",
+    "pretty-format": "^26.6.2",
     "ts-loader": "^8.0.11",
+    "ts-node": "^9.0.0",
     "typescript": "^4.1.2",
     "webpack": "^5.6.0",
     "webpack-cli": "^4.2.0",
