+++ collaboredit/src/frontend/components/home.tsx
