+++ collaboredit/src/frontend/components/research.tsx
