+++ collaboredit/src/frontend/components/simpleTextEdit.tsx
@@ -1,10 +1,10 @@
 import * as React from 'react';
-import { test_automerge, test_wss } from '../rtc/server';
+import { test_automerge } from '../../rtc/capture';
 import { Box } from 'rebass';
 import { Label, Textarea } from '@rebass/forms'; 
 
 const SimpleEditor = (props: any) => {
-	React.useEffect(() => { test_automerge(); test_wss();});
+	React.useEffect(() => { test_automerge(); });
 	return(
 <div>
 <Box p={5}
