+++ collaboredit/src/rtc/serverPanel.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 
-const serverPanel = (props) => {
+const serverPanel = (props: any) => {
 	return(
 	<div>
 		<p> server panel to go here </p>
